Debug update to resolve collision mesh glitch: Added printing of player position and mesh position in debug mode. Effectively removed reliance on friction; instant stopping

+++ src/net/piemaster/jario/systems/CollisionMeshSystem.java
@@ -32,7 +32,7 @@ public class CollisionMeshSystem extends EntityProcessingSystem
 		if(mesh != null)
 		{
 			Transform t = transformMapper.get(e);
-			meshMapper.get(e).setLocation(t.getX(), t.getY());
+			meshMapper.get(e).setLocation((int)t.getX(), (int)t.getY());
 		}
 	}
 }
